requirements.txt
llama-index==0.10.7
llama-index-readers-file==0.1.33
python-dotenv==1.0.0
pypdf>=4.0.1
streamlit==1.29.0
chromadb==0.4.22
openai>=1.1.0 